Re-enable tests

+++ programs/steward/src/score.rs
@@ -147,10 +147,7 @@ pub fn validator_score(
 
     let historical_commission_max = validator
         .history
-        .commission_range(
-            VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH as u16,
-            current_epoch as u16,
-        )
+        .commission_range(VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH as u16, current_epoch)
         .iter()
         .filter_map(|&i| i)
         .max()
